Rename tab components
