--- src/views/AreaView/components/AreaTab/index.js
@@ -1,5 +1,0 @@
-import { withStyles } from '@material-ui/core';
-import AreaTab from './AreaTab';
-import styles from '../../styles';
-
-export default withStyles(styles)(AreaTab);
