+++ src/views/AreaView/components/GeographicalTab/index.js
@@ -1,6 +1,6 @@
 import { withStyles } from '@material-ui/core';
 import { injectIntl } from 'react-intl';
-import UnitTab from './UnitTab';
+import GeographicalTab from './GeographicalTab';
 import styles from '../../styles';
 
-export default injectIntl(withStyles(styles)(UnitTab));
+export default injectIntl(withStyles(styles)(GeographicalTab));
