+++ src/views/AreaView/components/ServiceTab/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import ServiceTab from './ServiceTab';
+import styles from '../../styles';
+
+export default withStyles(styles)(ServiceTab);
